% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\encoding{UTF-8}
\name{agregar_dbs}
\alias{agregar_dbs}
\title{Función que permite añadir datos nuevos}
\usage{
agregar_dbs(base, ...)
}
\arguments{
\item{base}{base de datos data.frame o data.table de base, a la cual se le añadirá nueva información.}

\item{...}{argumentos pasados a las funciones \link{homologar_db}, \link{consolidar_db} y \link{depurar_ruts}.}
}
\description{
Esta función es un parser para las funciones individuales para agregar nuevos datos, de una manera más cómoda y fácil.
}
