% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\encoding{UTF-8}
\name{buscar_ruts}
\alias{buscar_ruts}
\title{Búsqueda por rut}
\usage{
buscar_ruts(
  ruts,
  db,
  cols = c("proveedor_rut", "proveedor_nombre_fantasia", "proveedor_mercado_nombre")
)
}
\arguments{
\item{ruts}{vector con los ruts a verificar.}

\item{db}{objeto dataframe de data.table.}

\item{cols}{columnas donde mirar.}
}
\description{
Búsqueda por rut
}
