% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{calcular_metricas}
\alias{calcular_metricas}
\title{Calcular métricas}
\usage{
calcular_metricas(x, cuantiles, IQR, coef)
}
\arguments{
\item{x}{vector númerico.}

\item{cuantiles}{vector númerico con los cuantiles deseados (valor entre 0 y 1).}

\item{IQR}{vector con el rango intercuantil de la forma c(min, max). Ambos valores deben estar en \code{cuantiles}.}

\item{coef}{coeficiente de dispersión (de castigo), usualmente 1.5.}
}
\description{
Calcular métricas
}
