% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computar.R
\name{computar_reclamos}
\alias{computar_reclamos}
\title{Procesar reclamos}
\usage{
computar_reclamos(
  mdatai,
  byCategoria = c("cierre_corto", "proveedor_mercado_nombre",
    "proveedor_mercado_categoria_nombre", "proveedor_rut", "cut_region"),
  byClase = c(industria = "proveedor_mercado_categoria_nombre", proveedor =
    "proveedor_rut", producto = "mercado_tipo_producto_nombre"),
  lags = 1:6,
  cuantiles = c(0, 0.25, 0.5, 0.75, 1),
  IQR = c(0.25, 0.75),
  coef = 1.5,
  n_min = 50,
  rango_fechas = c(NULL, NULL),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{mdatai}{data.frame con la base de datos.}

\item{byCategoria}{vector de strings con las variables consideradas como categorías.}

\item{byClase}{vector de strings con nombres (\link{setNames}), que incluya las variables de clase (categorización). El orden es importante.}

\item{lags}{vector númerico}

\item{cuantiles}{vector númerico con los cuantiles deseados (valor entre 0 y 1).}

\item{IQR}{vector con el rango intercuantil de la forma c(min, max). Ambos valores deben estar en \code{cuantiles}.}

\item{coef}{coeficiente de dispersión (de castigo), usualmente 1.5.}

\item{n_min}{número mínimo de observaciones}

\item{rango_fechas}{vector de la forma \code{c(min, max)} que incluye las fechas inferior y superior a considerar. Puede ser en formato fecha o
como string en formato YYYY-mm-dd.}

\item{verbose}{si es TRUE, devuelve mensajes sobre el avance del proceso, de lo contrario, no entrega ningún mensaje. Por defecto es TRUE.}

\item{...}{argumentos extras pasados a \link{agg_by}, como \code{usar_tiempo}.}
}
\description{
Procesar reclamos
}
