% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\encoding{UTF-8}
\name{consolidar_db}
\alias{consolidar_db}
\title{Consolidar base de datos}
\usage{
consolidar_db(..., db, grabar = FALSE)
}
\arguments{
\item{...}{base o bases de datos a agregar, previamente homologadas con \link{homologar_db}.

#' @importFrom data.table rbindlist}

\item{db}{db previamente homologada con \link{homologar_db}.}

\item{grabar}{si es TRUE, la unión de las bases de datos se guarda como un nuevo archivo actualizado.
Se recomienda utilizarlo sólo cuando se esté con certeza de que los valores no cambiarán. Por defecto es FALSE.}
}
\description{
Agrega nuevos registros a la base de datos de inicio.
}
