% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\encoding{UTF-8}
\name{depurar_ruts}
\alias{depurar_ruts}
\title{Homologa los ruts y su nombre de fantasía (dentro de lo posible).}
\usage{
depurar_ruts(db, tolerancia = 20, criterio = "reciente")
}
\arguments{
\item{db}{db previamente homologada con \link{homologar_db}.}

\item{tolerancia}{número de nombres de fantasía por rut máximo para asignar por mayoría, antes de considerar el registro como extraño.
Ejemplo, cuando no se sabe el rut de la empresa y se rellena con 111111, etc.}

\item{criterio}{citerio de asignación del nombre de fantasía al rut. Valores posibles son 'mayoria' o 'reciente'. Por defecto 'reciente'.}
}
\description{
Homologa los ruts y su nombre de fantasía (dentro de lo posible).
}
