% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\encoding{UTF-8}
\name{exportar}
\alias{exportar}
\title{Exportar diferentes elementos de la clase reclamos a csv.}
\usage{
exportar(obj, que = "ranking", donde = NULL, ...)
}
\arguments{
\item{obj}{objeto de la clase reclamos.}

\item{que}{vector de objetos a exportar. Opciones válidas son: \code{ranking}, \code{agrupados}.}

\item{donde}{ruta de salida. Si ninguna es específicada se usará el directorio actual de trabajo.}

\item{...}{parámetros adicionales para \link[data.table]{fwrite}, como separador de columnas, marcador de decimales, verbose, etc.}
}
\description{
Exportar diferentes elementos de la clase reclamos a csv.
}
\details{
Por completar, con la descripción de las columnas de cada output.
}
