% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computar.R
\encoding{UTF-8}
\name{generar_metricas}
\alias{generar_metricas}
\title{Genera metricas internas}
\usage{
generar_metricas(x, cuantiles, IQR, coef, byClase, exprs1, metricsN)
}
\arguments{
\item{x}{data.table data frame.}

\item{cuantiles}{vector númerico con los cuantiles deseados (valor entre 0 y 1).}

\item{IQR}{vector con el rango intercuantil de la forma c(min, max). Ambos valores deben estar en \code{cuantiles}.}

\item{coef}{coeficiente de dispersión (de castigo), usualmente 1.5.}

\item{byClase}{vector de strings con nombres (\link{setNames}), que incluya las variables de clase (categorización). El orden es importante.}

\item{exprs1}{expresión.}

\item{metricsN}{data.table data frame}
}
\description{
Subfuncón sin ninguna utilidad fuera del entorno interno, para estrucutrar de mejor manera los cálculos.
}
