% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{get_maj}
\alias{get_maj}
\title{Obtener registro con mayoría}
\usage{
get_maj(db, cols, bycol)
}
\arguments{
\item{db}{data.table dataframe}

\item{cols}{columna(s) de agrupación, por las que filtrarán los registros.}

\item{bycol}{columna(s) que contiene el registro que seerá usadao para buscar el mayor N.}
}
\value{

}
\description{
Obtener registro con mayoría
}
\examples{
1+1
}
