% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\name{homologar_arbol}
\alias{homologar_arbol}
\title{Homologa el árbol de mercados}
\usage{
homologar_arbol(arbol_motivo_legal)
}
\arguments{
\item{arbol_motivo_legal}{}
}
\value{

}
\description{
Permite realizar homologaciones y rellenar valores faltantes en las categorías de daño y facilidad de prueba
}
