% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\name{homologar_sii}
\alias{homologar_sii}
\title{Homologa datos del SII}
\usage{
homologar_sii(siiData)
}
\arguments{
\item{siiData}{data.frame de data.table con la información}
}
\value{

}
\description{
Permite homologar nombres de comunas y revisar algunas incosistencias
}
