% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\encoding{UTF-8}
\name{norm_rank}
\alias{norm_rank}
\title{Ranking normalizado}
\usage{
norm_rank(x)
}
\arguments{
\item{x}{un vector}
}
\value{
una lista de valores centrados y escalados y del cuantil al que corresponde cada observación.
}
\description{
Ranking normalizado
}
\examples{
norm_rank(1:10)
}
