% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{obtener_outliers}
\alias{obtener_outliers}
\title{Marca outliers}
\usage{
obtener_outliers(x, t_min, t_max, to_data_table = TRUE)
}
\arguments{
\item{x}{vector númerico}

\item{t_min}{límite inferior}

\item{t_max}{límite superior}

\item{to_data_table}{si es TRUE, transforma el ooutput a una data.table dataframe. Caso contrario, retorna una lista. Por defecto es TRUE.}
}
\value{

}
\description{
Marca outliers
}
