% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\encoding{UTF-8}
\name{obtener_reciente}
\alias{obtener_reciente}
\title{Obtener registro más reciente}
\usage{
obtener_reciente(db, cols, bycol)
}
\arguments{
\item{db}{data.table dataframe}

\item{cols}{columna(s) de agrupación, por las que filtrarán los registros.}

\item{bycol}{columna(s) que contiene el registro que seerá usadao para buscar el mayor N.}
}
\value{

}
\description{
Obtener registro más reciente
}
\details{
Basado en \code{caso_creacion_fecha}.
}
\examples{
1+1
}
