% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotme}
\alias{plotme}
\title{Gráfico de la serie}
\usage{
plotme(
  db,
  t_observado = "min",
  pos = 1,
  ascendente = TRUE,
  clase = "industria",
  metrics = NULL
)
}
\arguments{
\item{db}{objeto retornado por \link{computar_reclamos}.}

\item{t_observado}{periodo observado}

\item{pos}{posición del ranking a observar}

\item{ascendente}{se muestren de menor a mayor?}

\item{clase}{nombre de la clase que se desa visualziar. Debe corresponder al nombre asignado a \code{byClase} en \link{computar_reclamos}.}

\item{metrics}{métricas a visualizar. Si no se pasa argumento, se usarán todas.}
}
\description{
Gráfico de la serie
}
\examples{
1+1
}
