% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{print.reclamos}
\alias{print.reclamos}
\title{Print method para la clase reclamos}
\usage{
\method{print}{reclamos}(x, ...)
}
\arguments{
\item{x}{objeto de la clae \code{reclamos}}
}
\description{
Print method para la clase reclamos
}
