% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\encoding{UTF-8}
\name{seleccionar_ranking}
\alias{seleccionar_ranking}
\title{Seleccionar por ranking}
\usage{
seleccionar_ranking(dat, topn, nmin, reporte = TRUE, ...)
}
\arguments{
\item{dat}{objeto de la clase \code{reclamos}.}

\item{topn}{número de posiciones del ranking a incluir. Se pueden retornar más o menos valores,
dependiendo de si existen empates o el número final de registros.}

\item{nmin}{número mínimo de datos requeridos para estar presente en la selección.}

\item{reporte}{si es TRUE, imprime un reporte resumen.}

\item{...}{argumentos extras pasados a \link{rankear}.}
}
\description{
Seleccionar por ranking
}
\examples{
1+1
}
