% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot.R
\encoding{UTF-8}
\name{shinyPlot}
\alias{shinyPlot}
\title{Genera un dashboard con shiny}
\usage{
shinyPlot(obj)
}
\arguments{
\item{obj}{Objeto de la clase reclamos.}
}
\description{
Genera un dashboard con shiny
}
