% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\encoding{UTF-8}
\name{ver_ranking}
\alias{ver_ranking}
\title{Visualizar rankings}
\usage{
ver_ranking(
  x,
  t_observado = NULL,
  pos = 1,
  ascendente = NULL,
  clase = "industria"
)
}
\arguments{
\item{x}{objeto retornado por \link{computar_reclamos}.}

\item{t_observado}{periodo observado}

\item{pos}{posición del ranking a observar}

\item{ascendente}{se muestren de menor a mayor?}

\item{clase}{nombre de la clase que se desa visualziar. Debe corresponder al nombre asignado a \code{byClase} en \link{computar_reclamos}.}
}
\description{
Visualizar rankings
}
