% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\encoding{UTF-8}
\name{extraer_series}
\alias{extraer_series}
\title{EXtrae las series}
\usage{
extraer_series(obj)
}
\arguments{
\item{reclamos}{objeto de la clase \code{reclamos_seleccion}.}
}
\value{

}
\description{
EXtrae las series
}
