% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\encoding{UTF-8}
\name{homologar_db}
\alias{homologar_db}
\title{Homologa datos de entrada al formato apropiado}
\usage{
homologar_db(
  db,
  diccionario_columnas = NULL,
  codigos_comunales = NULL,
  arbol_motivo_legal = NULL,
  datos_sii = NULL,
  verbose = TRUE,
  full_output = TRUE
)
}
\arguments{
\item{db}{ruta al archivo, en excel (formato especial flexible).}

\item{diccionario_columnas}{ruta al archivo de diccionario de columnas en excel (formato especial estricto).}

\item{codigos_comunales}{ruta al archivo de codigos y nombres comunales en excel (formato especial estricto). Si 'skip', se salta este paso.}

\item{arbol_motivo_legal}{ruta al archivo de equivalencias de motivos legales en excel (formato especial estricto).}

\item{datos_sii}{ruta al archivo de datos del SII, en csv (formato especial estricto).}

\item{verbose}{si TRUE, imprime estados de avance.}

\item{full_output}{si TRUE, entrega todas las tablas usadas. Si FALSE, entrega solo la tabla principal modificada.}
}
\value{

}
\description{
Permite homologar nombres de columnas, comunas, motivos legales; también selecciona columnas de interés,
y añade datos del SII. La tabla queda lista para ser añadida a la tabla general de datos. La función trata de ser lo más
inteligente posible, pero es necesario mantener los formatos apropiados.
}
