% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\encoding{UTF-8}
\name{rankear}
\alias{rankear}
\title{Generar ranking}
\usage{
rankear(dt, col, t_observado = "max")
}
\arguments{
\item{dt}{data.table.}

\item{col}{nombre de la columna a evaluar.}

\item{t_observado}{tiempo en el cual se evalua el ranking.}
}
\description{
Generar ranking
}
\examples{
1+1
}
