% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot.R
\encoding{UTF-8}
\name{shinyPlot}
\alias{shinyPlot}
\title{Genera un dashboard con shiny}
\usage{
shinyPlot(obj, port = 6742)
}
\arguments{
\item{obj}{Objeto de la clase reclamos.}

\item{port}{puerto en el que se ejecuta el servidor web de la aplicación}
}
\description{
Levanta un dashboard de shiny, en un servidor web local que permite visualizar las alertas de manera interactiva.
}
